Refactor scripts into their own folder for organization purposes, and then refactor their style
