+++ Scripts/Simulation/ReferenceFrames/CoordinateSpace.cs
@@ -1,0 +1,22 @@
+﻿namespace GodotPrototype.Scripts.Simulation.ReferenceFrames;
+
+public enum CoordinateSpace
+{
+    RenderSpace = 0,
+    GalaxySpace = 1,
+    StarSpace = 2,
+    PlanetSpace = 3,
+    MoonSpace = 4,
+};
+public static class Extensions
+{
+    public static CoordinateSpace Increment(this CoordinateSpace coordLayer)
+    {
+        return (CoordinateSpace)((int)coordLayer + 1);
+    }
+
+    public static float GetConversionFactor(this CoordinateSpace from, CoordinateSpace to)
+    {
+        return GlobalValues.GetRefConversionFactor(from, to);
+    }
+}
