+++ project.godot
@@ -11,7 +11,8 @@ config_version=5
 [application]
 
 config/name="Godot-Prototype"
-config/features=PackedStringArray("4.3", "Forward Plus")
+run/main_scene="res://node_3d.tscn"
+config/features=PackedStringArray("4.3", "C#", "Forward Plus")
 config/icon="res://icon.svg"
 
 [dotnet]
